requirements.txt
asgiref==3.8.1
astroid==3.3.11
cffi==1.17.1
colorama==0.4.6
cryptography==45.0.2
dill==0.4.0
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
isort==6.0.1
jwt==1.3.1
mccabe==0.7.0
platformdirs==4.3.8
pycparser==2.22
PyJWT==2.9.0
pylint==3.3.7
sqlparse==0.5.3
tomlkit==0.13.3
tzdata==2025.2
